core: Resolve C4267 warning on MSVC

+++ src/core/dumping/backend.cpp
@@ -8,7 +8,8 @@
 namespace VideoDumper {
 
 VideoFrame::VideoFrame(std::size_t width_, std::size_t height_, u8* data_)
-    : width(width_), height(height_), stride(width * 4), data(data_, data_ + width * height * 4) {}
+    : width(width_), height(height_), stride(static_cast<u32>(width * 4)),
+      data(data_, data_ + width * height * 4) {}
 
 Backend::~Backend() = default;
 NullBackend::~NullBackend() = default;
